backend/requirements.txt
fastapi
uvicorn
playwright
openai
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
python-multipart
pytest
email-validator
